fix: simple refactor (indents and logs) and rename Nav to Home
