+++ components/GoogleSearch.js
@@ -3,8 +3,6 @@ import { Text, Icon } from 'native-base'
 
 var { GooglePlacesAutocomplete } = require('react-native-google-places-autocomplete');
 
-const homePlace = {description: 'Home', geometry: { location: { lat: 48.8152937, lng: 2.4597668 } }};
-const workPlace = {description: 'Work', geometry: { location: { lat: 48.8496818, lng: 2.2940881 } }};
 const fullStack = {description: 'FullStack', geometry: { location: { lat: 40.704980, lng: -74.009133 } }};
 
 class GoogleSearch extends Component {
