+++ index.ios.js
@@ -2,13 +2,13 @@ console.disableYellowBox = true;
 
 import React from 'react';
 import { AppRegistry } from 'react-native';
-import Nav from './Nav';
+import Home from './Home';
 
 
 export default class maptest extends React.Component {
     render() {
         return (
-            <Nav />
+            <Home />
         );
     }
 }
